+++ config/settings/base.py
@@ -113,6 +113,7 @@ TEMPLATES = [
                 "django.template.context_processors.i18n",
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
+                "apps.seo.context_processors.seo",
             ],
         },
     },